--- app/dashboard/Main.css
@@ -1,19 +1,0 @@
-/* Main.css */
-.fade-enter {
-    opacity: 0.01;
-  }
-  
-  .fade-enter-active {
-    opacity: 1;
-    transition: opacity 300ms ease-in;
-  }
-  
-  .fade-exit {
-    opacity: 1;
-  }
-  
-  .fade-exit-active {
-    opacity: 0.01;
-    transition: opacity 300ms ease-in;
-  }
-
