+++ app/dashboard/layout.tsx
@@ -10,9 +10,9 @@ export default function RootLayout({
     return (
         <Box className="w-screen bg-[#F6F6F6] md:p-[44px] sm:p-0  h-screen overflow-y-hidden ">
             <Header/>
-            <Box className="flex w-full md:flex-row flex-col-reverse ">
+            <Box className="flex w-full md:flex-row flex-col-reverse overflow-y-hidden">
             <SideBar/>
-            <Box className="h-[79vh] p-3 w-full md:h-screen overflow-y-auto">
+            <Box className="h-[79vh] p-3 w-full md:h-screen overflow-y-hidden">
             {children}
             </Box>
